cleanup redundant report controllers and redundant actions, reimplement login button in mobile view
