+++ app/templates/includes/_menu.html.twig
@@ -65,7 +65,6 @@
                             <li><a class="dropdown-item py-1" href="{{ path('admin_feedback_index') }}"><i class="fas fa-comments me-2 text-muted"></i>Feedback</a></li>
                             <li><a class="dropdown-item py-1" href="{{ path('app_news_index') }}"><i class="fas fa-newspaper me-2 text-muted"></i>Neuigkeiten Management</a></li>
                             <li><a class="dropdown-item py-1" href="{{ path('admin_consistency_index') }}"><i class="fas fa-search me-2 text-muted"></i>Datenkonsistenz</a></li>
-                            <li><a class="dropdown-item py-1" href="{{ path('app_report_list') }}"><i class="fas fa-chart-bar me-2 text-muted"></i>Report Templates</a></li>
                             <li><hr class="dropdown-divider"></li>
                             <li><h6 class="dropdown-header text-warning"><i class="fas fa-link me-1"></i>Zuweisungen</h6></li>
                             <li><a class="dropdown-item py-1" href="{{ path('admin_users_index') }}"><i class="fas fa-user-cog me-2 text-muted"></i>Benutzer</a></li>
@@ -98,6 +97,14 @@
                         <button class="nav-link text-danger btn btn-navbar-logout w-100 text-start justify-content-start" id="button_logout">
                             <i class="fas fa-sign-out-alt me-2"></i>
                             Abmelden
+                        </button>
+                    </li>
+                {% else %}
+                    {# Login Button für Gäste #}
+                    <li class="nav-item">
+                        <button class="btn btn-light" onclick="const modalManager = (window.modalManager || window.ModalManager); modalManager.showModal('auth_modal')">
+                            <i class="fas fa-sign-in-alt me-1"></i>
+                            Anmelden
                         </button>
                     </li>
                 {% endif %}
